+++ src/index.ts
@@ -22,6 +22,7 @@ export * from "./errors/transaction-rollback.error";
 export * from "./errors/unsafe-filter.error";
 export * from "./errors/unsafe-raw-expression.error";
 export * from "./errors/unsupported-lean-operation.error";
+export * from "./errors/unsupported-query-operation.error";
 export * from "./errors/unsupported-update-operation.error";
 
 // Core Services